Add props for accessing parent blocks & chapters
Every entry that is not root can now return the chapter it belongs to
via the property `chapter`.
Every entry that is not root or chapter can now return the closest
block it belongs to via the property `block`.
The properties accessing the children (e.g. `chapters` & `blocks`)
were renamed to be in plural, since any entry can only have one parent,
but multiple children.